Added fridge repository interface
