+++ app/src/main/java/com/example/crisisfridge/data/database/dao/PantryDao.java
@@ -4,13 +4,13 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.example.crisisfridge.data.inventory.dao;
+package com.example.crisisfridge.data.database.dao;
 
 
 import androidx.room.Dao;
 
+
 @Dao
-public interface RecipeDao {
-
-
+public interface PantryDao {
 }
+
