+++ app/src/main/java/com/example/crisisfridge/data/database/dao/RecipeIngredientsDao.java
@@ -4,7 +4,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.example.crisisfridge.data.inventory.dao;
+package com.example.crisisfridge.data.database.dao;
 
 import androidx.room.Dao;
