+++ app/src/main/java/com/example/crisisfridge/data/database/dao/ShoppingListDao.java
@@ -4,16 +4,15 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.example.crisisfridge.data.inventory.dao;
-
-
+package com.example.crisisfridge.data.database.dao;
 
 
 import androidx.room.Dao;
 
 
-
 @Dao
-public interface PantryDao {
+public interface ShoppingListDao {
+
+
 }
