+++ app/src/main/java/com/example/crisisfridge/data/database/entity/FridgeItemEntity.java
@@ -1,35 +1,33 @@
-package com.example.crisisfridge.data.inventory.entity;
+package com.example.crisisfridge.data.database.entity;
 
 import androidx.room.Entity;
 import androidx.room.ForeignKey;
 import androidx.room.Index;
 import androidx.room.PrimaryKey;
 
-import com.example.crisisfridge.data.model.inventory.interfaces.ShoppingList;
+import com.example.crisisfridge.data.model.dataModel.FridgeItem;
 
-import java.util.Objects;
-
-@Entity(tableName = "shopping_list",
+@Entity(tableName = "fridge",
         foreignKeys = @ForeignKey(
                 entity = ProductTypeEntity.class,
                 parentColumns = "id",
-                childColumns = "productTypeId",
+                childColumns = "productId",
                 onDelete = ForeignKey.CASCADE),
-        indices = @Index("productTypeId")
+        indices = @Index("productId")
 )
-public class ShoppingListEntity implements ShoppingList {
+public class FridgeItemEntity implements FridgeItem {
 
     @PrimaryKey
     private int id;
-    private int productTypeId;
+    private int productId;
     private float quantity;
-    private boolean bought;
+    private long expirationDate;
 
-    public ShoppingListEntity(int id, int productTypeId, float quantity, boolean bought) {
+    public FridgeItemEntity(int id, int productId, float quantity, long expirationDate) {
         this.id = id;
-        this.productTypeId = productTypeId;
+        this.productId = productId;
         this.quantity = quantity;
-        this.bought = bought;
+        this.expirationDate = expirationDate;
     }
 
     @Override
@@ -38,8 +36,8 @@ public class ShoppingListEntity implements ShoppingList {
     }
 
     @Override
-    public int getProductTypeId() {
-        return productTypeId;
+    public int getProductId() {
+        return productId;
     }
 
     @Override
@@ -48,8 +46,8 @@ public class ShoppingListEntity implements ShoppingList {
     }
 
     @Override
-    public boolean isBought() {
-        return bought;
+    public long getExpirationDate() {
+        return expirationDate;
     }
 
     @Override
@@ -57,20 +55,20 @@ public class ShoppingListEntity implements ShoppingList {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
 
-        ShoppingListEntity that = (ShoppingListEntity) o;
+        FridgeItemEntity that = (FridgeItemEntity) o;
 
         if (id != that.id) return false;
-        if (productTypeId != that.productTypeId) return false;
+        if (productId != that.productId) return false;
         if (Float.compare(that.quantity, quantity) != 0) return false;
-        return bought == that.bought;
+        return expirationDate == that.expirationDate;
     }
 
     @Override
     public int hashCode() {
         int result = id;
-        result = 31 * result + productTypeId;
+        result = 31 * result + productId;
         result = 31 * result + (quantity != +0.0f ? Float.floatToIntBits(quantity) : 0);
-        result = 31 * result + (bought ? 1 : 0);
+        result = 31 * result + (int) (expirationDate ^ (expirationDate >>> 32));
         return result;
     }
 }
